refactor: rename world permissions manager
