+++ src/components.ts
@@ -19,7 +19,7 @@ import { createPlayerStorageComponent } from './adapters/player-storage'
 import { createWorldStorageComponent } from './adapters/world-storage'
 import { createWorldsContentServerComponent } from './adapters/worlds-content-server'
 import { getDbConnectionString } from './logic/utils'
-import { createWorldPermissionComponent } from './logic/world-permission'
+import { createWorldPermissionsManagerComponent } from './logic/world-permissions-manager'
 import { metricDeclarations } from './metrics'
 import type { AppComponents, GlobalContext } from './types'
 
@@ -62,7 +62,7 @@ export async function initComponents(): Promise<AppComponents> {
   const playerStorage = createPlayerStorageComponent({ pg })
   const envStorage = createEnvStorageComponent({ pg, encryption })
   const worldsContentServer = await createWorldsContentServerComponent({ fetcher, config })
-  const worldPermission = createWorldPermissionComponent({ worldsContentServer })
+  const worldPermissionsManager = createWorldPermissionsManagerComponent({ worldsContentServer })
 
   return {
     fetcher,
@@ -77,7 +77,7 @@ export async function initComponents(): Promise<AppComponents> {
     playerStorage,
     envStorage,
     worldsContentServer,
-    worldPermission,
+    worldPermissionsManager,
     schemaValidator
   }
 }
